+++ src/config/defaults.ts
@@ -15,10 +15,15 @@ import {
   EnqueueUrlOptions,
   SpiderOptions,
 } from '../spider/index.js';
-import { PageDataOptions, PageContentOptions } from '../tools/html/index.js';
+import {
+  PageDataOptions,
+  PageContentOptions,
+  HtmlToTextOptions,
+} from '../tools/html/index.js';
 import { PageTechnologyOptions } from '../tools/browser/index.js';
 import { readPackageUpSync } from 'read-pkg-up';
 import { AqQuery } from 'aql-builder';
+import { PageAnalysisOptions } from '../tools/analyze-page.js';
 
 export const urlNormalizerDefaults: NormalizerOptions = {
   forceProtocol: 'https:',
@@ -34,7 +39,7 @@ export const urlNormalizerDefaults: NormalizerOptions = {
 
 export const urlDiscoveryDefaults: EnqueueUrlOptions = {
   limit: 1_000_000,
-  selector: 'body a',
+  selectors: 'body a',
   save: UrlMatchStrategy.All,
   enqueue: UrlMatchStrategy.SameDomain,
   prioritize: false,
@@ -42,11 +47,16 @@ export const urlDiscoveryDefaults: EnqueueUrlOptions = {
   respectRobots: false,
   checkSitemaps: false,
   prioritizeSitemaps: false,
-  discardEmptyLinks: true,
-  discardAnchorOnlyLinks: true,
-  discardNonWebLinks: false,
-  discardUnparsableLinks: false,
-  discardExistingLinks: true,
+  discardEmpty: true,
+  discardlocalAnchors: true,
+  discardNonWeb: false,
+  discardUnparsable: false,
+  discardExisting: true,
+};
+
+export const htmlToTextDefaults: HtmlToTextOptions = {
+  wordwrap: false,
+  selectors: [{ selector: 'a', options: { ignoreHref: true } }],
 };
 
 export const spiderDefaults: Partial<SpiderOptions> = {
@@ -99,16 +109,34 @@ export const defaultQueries: Record<string, AqQuery> = {
   },
 };
 
+export const arangoDefaults = {
+  databaseName: 'spidergram',
+  url: 'http://127.0.0.1:8529',
+  auth: {
+    username: 'root',
+    password: '',
+  },
+};
+
+export const analyzePageDefaults: PageAnalysisOptions = {
+  data: pageDataDefaults,
+  content: {},
+  tech: {},
+  regions: {},
+  propertyMap: {},
+};
+
 export const spidergramDefaults: SpidergramConfig = {
   debug: false,
   logLevel: 'error',
   storageDirectory: path.join(process.cwd(), 'storage'),
-  arango: { databaseName: 'spidergram' },
+  arango: arangoDefaults,
   crawlee: {},
   spider: spiderDefaults,
-  htmlToText: {},
+  htmlToText: htmlToTextDefaults,
   queries: defaultQueries,
   reports: {},
+  pageAnalysis: analyzePageDefaults,
   pageData: pageDataDefaults,
   pageContent: pageContentDefaults,
   pageTechnologies: pageTechnologyDefaults,