+++ src/spider/links/index.ts
@@ -1,6 +1,6 @@
 export * from './enqueue-options.js';
 export * from './enqueue-wrapper.js';
 export * from './find-urls.js';
-export * from './filter-urls.js';
+export * from './filter-url.js';
 export * from './save-urls.js';
 export * from './enqueue-requests.js';