+++ src/spider/spider-options.ts
@@ -2,6 +2,7 @@ import { Dictionary, LogLevel, PlaywrightCrawlerOptions } from 'crawlee';
 import { SpiderHook } from './hooks/index.js';
 import { EnqueueUrlOptions } from './links/index.js';
 import { SpiderRequestHandler } from './handlers/index.js';
+import { PageAnalysisOptions } from '../tools/analyze-page.js';
 
 export type SpiderOptions = InternalSpiderOptions &
   Omit<
@@ -41,6 +42,14 @@ export interface InternalSpiderOptions extends Dictionary {
    * @type {SpiderRequestHandler}
    */
   pageHandler?: SpiderRequestHandler;
+
+  /**
+   * Options to control the pageHandler's parsing of HTML content. This can be
+   * used to frontload certain analysis work (like technology fingerprinting) or
+   * delay time consuming operations (like named entity extraction) until a post-
+   * crawl analysis step.
+   */
+  pageOptions?: PageAnalysisOptions;
 
   /**
    * A dictionary of request handlers, keyed by the request label they're