+++ src/tools/index.ts
@@ -1,10 +1,12 @@
 export * as TextTools from './text/index.js';
 export * as HtmlTools from './html/index.js';
 export * as GoogleTools from './google/index.js';
-export * as PageUtils from './page.js';
 export * as HierarchyTools from './hierarchy/index.js';
 export * as BrowserTools from './browser/index.js';
 export * as FileTools from './file/index.js';
+export * from './find-property-value.js';
+export * from './analyze-page.js';
+export * from './relink-resource.js';
 
 export { Robots } from './robots.js';